added proper mixin for mobile sizes, variable for sidebar widths, removed loading spinner margin for better centering (added to message), credit tooltips now on bottom for readability, redirect button now grouped with user auth screen, credits icon stays lit while viewing to indicate one must click to get back to discord auth view, mobile friendly homepage
