+++ src/components/Auth/Auth.scss
@@ -17,6 +17,18 @@
   h2 {
     margin-bottom: 1rem;
   }
+
+  // So only applies within auth-user
+  .user-avatar-container {
+    transition: 0.2s;
+    
+      &:hover {
+        transform: translateY(-5px);
+        transform: scale(1.1);
+        box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+        cursor: pointer;
+      }
+  }
 }
   
 .user-avatar-container {
@@ -27,6 +39,10 @@
   margin-bottom: 2.5rem;
 
   .user-avatar {
+    position: absolute;
+    top: 0;
+    left: 0;
+
     width: 100px;
     height: 100px;
     border-radius: 50%;
@@ -61,6 +77,7 @@
 
 .auth-button {
   padding: 0.5rem 1rem;
+  margin-bottom: 1rem;
   font-size: 1rem;
   border: none;
   border-radius: 4px;
@@ -90,4 +107,33 @@
   text-align: center;
   font-style: italic;
   color: $light-grey-1;
+}
+
+.main-action-button {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  max-height: 60px;
+  margin-top: 2rem;
+  padding: 1rem 2rem;
+  font-size: 1.2rem;
+  border: none;
+  border-radius: 8px;
+  cursor: pointer;
+  transition: background-color 0.2s;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+  @include theme("background", "color-button");
+  @include theme("color", "color-button-text");
+
+  &:hover {
+    @include theme("background", "color-button-hover");
+  }
+
+  .loading-spinner {
+    filter: invert(1);
+  }
+
+  @include mobile {
+    display: none;
+  }
 }
