+++ src/components/Auth/Auth.tsx
@@ -1,4 +1,6 @@
-import React from "react";
+import React, { useState, useEffect } from "react";
+import { useNavigate } from "react-router-dom";
+import { API_URL } from "../../App";
 import LoadingSpinner from "../Loading/LoadingSpinner";
 import { useAuth } from "../../context/AuthContext";
 import { UserAvatar } from "../UserList/UserList";
@@ -10,9 +12,23 @@ interface AuthProps {
 
 const Auth: React.FC<AuthProps> = ({ onlyButtons = false }) => {
   const { user, loading, login, logout } = useAuth();
+  const navigate = useNavigate();
+
+  const handleButtonClick = () => {
+    if (!user) {
+      return;
+    }
+    else {
+      navigate(`/user/${user.id}`);
+    }
+  };
 
   if (loading) {
-    return <LoadingSpinner message="Loading user..." />;
+    return (
+      <div className="auth-container">
+        <LoadingSpinner message="Loading user..." />
+      </div>
+    );
   }
 
   if (onlyButtons) {
@@ -28,7 +44,7 @@ const Auth: React.FC<AuthProps> = ({ onlyButtons = false }) => {
   }
 
   return (
-    <div className="auth-container">
+    <div className="auth-container" onClick={handleButtonClick}>
       {user ? (
         <div className="auth-user">
           <h2>Welcome, {user.username}!</h2>
@@ -40,6 +56,58 @@ const Auth: React.FC<AuthProps> = ({ onlyButtons = false }) => {
           <button className="auth-button login-button" onClick={login}>Login with Discord</button>
         </div>
       )}
+    <RedirectButton />
+    </div>
+  );
+};
+
+const RedirectButton: React.FC = () => {
+  const [hasScores, setHasScores] = useState<boolean | null>(null);
+  const { user } = useAuth();
+  const navigate = useNavigate();
+
+  useEffect(() => {
+    if (user) {
+      fetch(`${API_URL}/api/user/${user.id}/has-scores`, {
+        headers: {
+          "Authorization": `Bearer ${localStorage.getItem("auth_token")}`
+        }
+      })
+        .then(response => response.json())
+        .then(data => setHasScores(data.has_scores))
+        .catch(error => {
+          console.error("Error checking user scores:", error);
+          setHasScores(null);
+        });
+    }
+  }, [user]);
+
+  const handleButtonClick = () => {
+    if (!user) {
+      return;
+    }
+    if (hasScores) {
+      navigate("/songs");
+    } else {
+      navigate(`/user/${user.id}`);
+    }
+  };
+
+  if (!user) {
+    return (
+      <div className="redirection-container"></div>
+    );
+  }
+
+  return (
+    <div className="redirection-container">
+      <button onClick={handleButtonClick} className="main-action-button">
+        {hasScores === null || hasScores ? (
+          "Check leaderboards here!"
+        ) : (
+          "Upload scores here!"
+        )}
+      </button>
     </div>
   );
 };
