+++ src/components/Extras/Credits.scss
@@ -22,6 +22,7 @@
       display: flex;
       align-items: center;
       padding: 1rem;
+      min-width: 220px;
       border-radius: 8px;
       text-decoration: none;
       transition: transform 0.2s, box-shadow 0.2s;
@@ -34,8 +35,6 @@
       }
 
       .avatar {
-        width: 60px;
-        height: 60px;
         border-radius: 50%;
         margin-right: 1rem;
       }
@@ -50,12 +49,37 @@
   }
 }
 
-@media (max-width: 768px) {
+@include mobile {
+  .credits {
+    h2 {
+      margin-top: 1rem;
+    }
+    .contributors-list {
+      .contributor {
+        min-height: 50px;
+        max-height: 50px;
+        // padding: 0;
+      }
+    }
+
+    .avatar {
+      width: 30px;
+      height: 30px;
+    }
+  }
+}
+
+@include tabletplus {
   .credits {
     .contributors-list {
       .contributor {
-        width: 100%;
+        min-width: 200px;
       }
+    }
+
+    .avatar {
+      width: 60px;
+      height: 60px;
     }
   }
 }
