+++ src/components/Loading/LoadingSpinner.scss
@@ -12,12 +12,12 @@
   .spinner-icon {
     width: 50px;
     height: 50px;
-    margin-bottom: 10px;
     @include theme("filter", "filter-invert");
   }
   
   .loading-message,
   .timeout-message {
+    margin-top: 10px;
     font-size: 1rem;
     @include theme("color", "color-text");
   }
