+++ src/styles/_mixins.scss
@@ -1,3 +1,33 @@
+$sidebar-width: 44px;
+$sidebar-width-expanded: 200px;
+
+$screen-mobile-max: 768px;
+$screen-tablet-max: 1200px;
+
+@mixin mobile {
+  @media (max-width: #{$screen-mobile-max - 1}) {
+    @content;
+  }
+}
+
+@mixin tablet {
+  @media (min-width: #{$screen-mobile-max}) and (max-width: #{$screen-tablet-max - 1}) {
+    @content;
+  }
+}
+
+@mixin tabletplus {
+  @media (min-width: #{$screen-mobile-max}) {
+    @content;
+  }
+}
+
+@mixin desktop {
+  @media (min-width: #{$screen-tablet-max}) {
+    @content;
+  }
+}
+
 @mixin theme($key, $color) {
   @each $theme-name, $theme-color in $themes {
       .theme-#{$theme-name} & {
